Fix a type issue found by CoreQL

+++ common/ffsparser.cpp
@@ -3901,7 +3901,7 @@ USTATUS FfsParser::parseIntelMicrocodeHeader(const UByteArray & microcode, const
             + (extendedHeader->Checksum == extendedCalculated ? UString("valid") : usprintf("invalid, should be %08Xh", extendedCalculated));
             
             const INTEL_MICROCODE_EXTENDED_HEADER_ENTRY* firstEntry = (const INTEL_MICROCODE_EXTENDED_HEADER_ENTRY*)(extendedHeader + 1);
-            for (UINT8 i = 0; i < extendedHeader->EntryCount; i++) {
+            for (UINT32 i = 0; i < extendedHeader->EntryCount; i++) {
                 const INTEL_MICROCODE_EXTENDED_HEADER_ENTRY* entry = (const INTEL_MICROCODE_EXTENDED_HEADER_ENTRY*)(firstEntry + i);
                 
                 // Recalculate checksum after patching
